Added ProvidedContextTrait and also use it in rules actions now.
