+++ src/Engine/RulesActionBase.php
@@ -8,11 +8,14 @@
 namespace Drupal\rules\Engine;
 
 use Drupal\Core\Plugin\ContextAwarePluginBase;
+use Drupal\rules\Context\ProvidedContextTrait;
 
 /**
  * Base class for rules actions.
  */
 abstract class RulesActionBase extends ContextAwarePluginBase implements RulesActionInterface {
+
+  use ProvidedContextTrait;
 
   /**
    * The plugin configuration.
