+++ src/Engine/RulesActionInterface.php
@@ -9,10 +9,11 @@ namespace Drupal\rules\Engine;
 
 use Drupal\Component\Plugin\ContextAwarePluginInterface;
 use Drupal\Core\Action\ActionInterface;
+use Drupal\rules\Context\ProvidedContextPluginInterface;
 
 /**
  * Extends the core ActionInterface to provide context.
  */
-interface RulesActionInterface extends ActionInterface, ContextAwarePluginInterface {
+interface RulesActionInterface extends ActionInterface, ContextAwarePluginInterface, ProvidedContextPluginInterface {
 
 }
