+++ src/Engine/RulesConditionBase.php
@@ -7,15 +7,16 @@
 
 namespace Drupal\rules\Engine;
 
-use Drupal\Component\Plugin\Exception\ContextException;
 use Drupal\Core\Executable\ExecutableManagerInterface;
-use Drupal\Core\Plugin\Context\Context;
 use Drupal\Core\Plugin\ContextAwarePluginBase;
+use Drupal\rules\Context\ProvidedContextTrait;
 
 /**
  * Base class for rules conditions.
  */
 abstract class RulesConditionBase extends ContextAwarePluginBase implements RulesConditionInterface {
+
+  use ProvidedContextTrait;
 
   /**
    * The condition manager to proxy execute calls through.
@@ -30,13 +31,6 @@ abstract class RulesConditionBase extends ContextAwarePluginBase implements Rule
    * @var array
    */
   protected $configuration;
-
-  /**
-   * The data objects that are provided by this condition.
-   *
-   * @var \Drupal\Component\Plugin\Context\ContextInterface[]
-   */
-  protected $provided;
 
   /**
    * {@inheritdoc}
@@ -119,44 +113,6 @@ abstract class RulesConditionBase extends ContextAwarePluginBase implements Rule
    */
   public function calculateDependencies() {
     return [];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setProvidedValue($name, $value) {
-    $this->getProvided($name)->setContextValue($value);
-    return $this;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getProvided($name) {
-    // Check for a valid context value.
-    if (!isset($this->provided[$name])) {
-      $this->provided[$name] = new Context($this->getProvidedDefinition($name));
-    }
-    return $this->provided[$name];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getProvidedDefinition($name) {
-    $definition = $this->getPluginDefinition();
-    if (empty($definition['provides'][$name])) {
-      throw new ContextException(sprintf("The %s provided context is not valid.", $name));
-    }
-    return $definition['provides'][$name];
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getProvidedDefinitions() {
-    $definition = $this->getPluginDefinition();
-    return !empty($definition['provides']) ? $definition['provides'] : array();
   }
 
 }
